Switched my site to use angularJS, putting here as reference
Switched my site to angularJS, making it a single page application (as
opposed to two pages before, but still). Locale switching is now
handled by toggling a single variable in controllers.js and the
appropriate value is set using a custom filter in apps.js.  To do:
figure out better scoping in the interval callback, animation stuff,
not hardcoding the strings in controller.js

Also note that if you run this on your local machine, it may not work
on Chrome because of cross-domain async request limiations on locally
run sites.  Should work in Firefox though